refactor: useGoToPage를 useHistory로 분리한다
